requirements.txt
requests~=2.32.3
tabulate~=0.9.0
PyYAML~=6.0.2
urllib3~=2.4.0
python-dotenv~=1.1.0